requirements.txt
alabaster==0.7.10
appdirs==1.4.3
appnope==0.1.0
astroid==1.5.2
autopep8==1.3.1
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
coverage==4.4.1
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.32.0
django-angular==0.8.4
django-bower==5.2.0
django-braces==1.11.0
django-ckeditor==5.2.2
django-datatables-view==1.14.0
django-debug-toolbar==1.8
django-enumfields==0.9.0
django-extensions==1.7.9
django-fastcgi==0.0.1
django-filter==1.0.3
django-libs==1.67.12
django-markdown-app==0.9.2
django-markdown-deux==1.0.5
django-mptt==0.8.7
django-pipeline==1.6.12
djangorestframework==3.6.3
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
gnureadline==6.3.3
google-api-python-client==1.6.2
googlemaps==2.4.6
gunicorn==19.7.1
httplib2==0.10.3
imagesize==0.7.1
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
Jinja2==2.9.6
Js2Py==0.44
lazy-object-proxy==1.3.1
linecache2==1.0.0
lxml==3.7.3
Markdown==2.6.8
markdown2==2.3.4
MarkupSafe==1.0
mccabe==0.6.1
mypy-lang==0.4.6
mysqlclient==1.3.10
oauth2client==4.1.0
oauthlib==2.0.2
packaging==16.8
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.4.0
ply==3.10
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
Pygments==2.2.0
pyjsparser==2.4.5
pylint==1.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
rsa==3.4.2
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
slimit==0.8.1
snowballstemmer==1.2.1
Sphinx==1.6.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
tox==2.7.0
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.1
tzlocal==1.4
Unidecode==0.4.20
unittest2==1.1.0
uritemplate==3.0.0
virtualenv==15.1.0
wcwidth==0.1.7
whitenoise==3.3.0
wrapt==1.10.10
yahoo-finance==1.4.0
yapf==0.16.1
